+++ clique.cc
