+++ clique.h
