+++ edge.cc
