+++ edge.h
