+++ fast_set.h
